// service/weather/api.ts
export interface Location {
  /** @example "New York" */
  name: string;
  /** @example "New York" */
  region: string;
  /** @example "United States of America" */
  country: string;
  /** @example 40.71 */
  lat: number;
  /** @example -74.01 */
  lon: number;
  /** @example "America/New_York" */
  tz_id: string;
  /**
   * @format int32
   * @example 1658522976
   */
  localtime_epoch: number;
  /** @example "2022-07-22 16:49" */
  localtime: string;
}

export interface Current {
  /**
   * @format int32
   * @example 1658522700
   */
  last_updated_epoch: number;
  /** @example "2022-07-22 16:45" */
  last_updated: string;
  /** @example 34.4 */
  temp_c: number;
  /** @example 93.9 */
  temp_f: number;
  /**
   * @format int32
   * @example 1
   */
  is_day: number;
  condition: {
    /** @example "Partly cloudy" */
    text: string;
    /** @example "//cdn.weatherapi.com/weather/64x64/day/116.png" */
    icon: string;
    /**
     * @format int32
     * @example 1003
     */
    code: number;
  };
  /** @example 16.1 */
  wind_mph: number;
  /** @example 25.9 */
  wind_kph: number;
  /** @example 180 */
  wind_degree: number;
  /** @example "S" */
  wind_dir: string;
  /** @example 1011 */
  pressure_mb: number;
  /** @example 29.85 */
  pressure_in: number;
  /** @example 0 */
  precip_mm: number;
  /** @example 0 */
  precip_in: number;
  /** @example 31 */
  humidity: number;
  /** @example 75 */
  cloud: number;
  /** @example 37 */
  feelslike_c: number;
  /** @example 98.6 */
  feelslike_f: number;
  /** @example 16 */
  vis_km: number;
  /** @example 9 */
  vis_miles: number;
  /**
   * @format int32
   * @example 8
   */
  uv: number;
  /** @example 11.6 */
  gust_mph: number;
  /** @example 18.7 */
  gust_kph: number;
  air_quality: {
    /** @example 293.70001220703125 */
    co: number;
    /** @example 18.5 */
    no2: number;
    /** @example 234.60000610351562 */
    o3: number;
    /** @example 12 */
    so2: number;
    /** @example 13.600000381469727 */
    pm2_5: number;
    /** @example 15 */
    pm10: number;
    /**
     * @format int32
     * @example 1
     */
    "us-epa-index": number;
    /**
     * @format int32
     * @example 2
     */
    "gb-defra-index": number;
  };
}